+++ lib/chatterbox/rails_catcher.rb
@@ -12,17 +12,13 @@ module Chatterbox
     # Overrides the rescue_action method in ActionController::Base, but does not inhibit
     # any custom processing that is defined with Rails 2's exception helpers.
     def rescue_action_in_public_with_chatterbox(exception)
-      Chatterbox.logger.debug { "Chatterbox caught exception #{exception} - about to handle" }
+      Chatterbox.logger.debug { "Chatterbox caught exception '#{exception}' - about to handle" }
       options = { :exception => exception }
       options.merge!(:request => request) if self.respond_to?(:request)
       Chatterbox::ExceptionNotification.handle(options)
-      Chatterbox.logger.debug { "Chatterbox handing exception #{exception} off to normal rescue handling" }
+      Chatterbox.logger.debug { "Chatterbox handing exception '#{exception}' off to normal rescue handling" }
       rescue_action_in_public_without_chatterbox(exception)
     end
     
-    private
-    
-    
   end
-  
 end